chore: refactor InMemoryWaitingRoomIdGenerator and InMemoryWaitingRoomRepository packages for better organization
