+++ src/main/java/com/icebreaker/be/infra/generator/InMemoryWaitingRoomIdGenerator.java
@@ -1,8 +1,9 @@
-package com.icebreaker.be.infra;
+package com.icebreaker.be.infra.generator;
 
 import com.icebreaker.be.domain.room.WaitingRoomIdGenerator;
-import java.util.UUID;
 import org.springframework.stereotype.Component;
+
+import java.util.UUID;
 
 @Component
 public class InMemoryWaitingRoomIdGenerator implements WaitingRoomIdGenerator {
