+++ src/main/java/com/icebreaker/be/infra/persistence/InMemoryWaitingRoomRepository.java
@@ -1,11 +1,12 @@
-package com.icebreaker.be.infra;
+package com.icebreaker.be.infra.persistence;
 
 import com.icebreaker.be.domain.room.WaitingRoom;
 import com.icebreaker.be.domain.room.WaitingRoomRepository;
+import org.springframework.stereotype.Repository;
+
 import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
-import org.springframework.stereotype.Repository;
 
 @Repository
 public class InMemoryWaitingRoomRepository implements WaitingRoomRepository {
